fix: address review comments
- Refactor ResolutionSession into its own file
- Ensure session.exit is called when an error is thrown
- Fix test names
